backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Werkzeug==2.3.7
Pillow==10.0.1
qrcode==7.4.2
python-dotenv==1.0.0
email-validator==2.1.0.post1
PyMySQL==1.1.0
cryptography==41.0.5
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
bcrypt==4.0.1
python-dateutil==2.8.2
requests==2.31.0 